Update dijkstras.js
Hey! I was using your JavaScript Dijkstras example as my AI pathing algorithm in a game I'm making. I was looking for performance optimizations and noticed that the algorithm was sorting the nodes list after every new item was enqueued, which didn't seem necessary in all cases. Proposing this change if you think it makes sense. Admittedly I'm not too familiar with the algorithm, and while it working in the cases I've tried I may have missed something. Cheers!